move streaming into separate folder
